Extract views & enums
